more route cleanup & security
