+++ hamza-client/src/modules/products/components/product-group-store/index.tsx
@@ -16,6 +16,7 @@ type Props = {
 
 const ProductCardGroup = ({ vendorName, filterByRating, category }: Props) => {
     // Get products from vendor
+    //TODO: MOVE TO INDEX.TS
     const { data, error, isLoading } = useQuery(
         ['products', { vendor: vendorName }],
         () => {
