+++ hamza-server/src/api/custom/checkout/route.ts
@@ -40,34 +40,6 @@ export const GET = async (req: MedusaRequest, res: MedusaResponse) => {
         handler.logger.debug(`returning checkout data: ${output}`);
         res.send({ orders: output });
     });
-
-    /*
-    const logger: Logger = req.scope.resolve('logger');
-    const { cart_id } = req.query;
-
-    try {
-        const orders = await orderService.getOrdersForCart(cart_id.toString());
-        const output: ICheckoutData[] = [];
-        orders.forEach((o) => {
-            output.push({
-                order_id: o.id,
-                cart_id: o.cart_id,
-                wallet_address: o.store?.owner?.wallet_address ?? '',
-                currency_code: o.payments[0].currency_code,
-                amount: o.payments[0].amount,
-                massmarket_amount: o.massmarket_amount,
-                massmarket_order_id: o.massmarket_order_id,
-                massmarket_ttl: o.massmarket_ttl,
-                orders,
-            });
-        });
-        logger.debug(`returning checkout data: ${output}`);
-        res.send({ orders: output });
-    } catch (e) {
-        logger.error(e);
-        res.send({ message: e.message });
-    }
-    */
 };
 
 export const POST = async (req: MedusaRequest, res: MedusaResponse) => {
