+++ hamza-server/src/api/custom/variant/count/route.ts
@@ -1,12 +1,12 @@
-import type { MedusaRequest, MedusaResponse, Logger } from '@medusajs/medusa';
+import type { MedusaRequest, MedusaResponse } from '@medusajs/medusa';
 import ProductVariantService from '../../../../services/product-variant';
-import { readRequestBody } from '../../../../utils/request-body';
 import { RouteHandler } from '../../../route-handler';
 
 export const POST = async (req: MedusaRequest, res: MedusaResponse) => {
     const productVariantService: ProductVariantService = req.scope.resolve(
         'productVariantService'
     );
+
     const handler: RouteHandler = new RouteHandler(req, res, 'POST', '/custom/variant/count', ['variant_id']);
 
     await handler.handle(async () => {
